refactor: fix the part that searches for the user name
There was an error when the user ran the script with sudo, it stored sudo instead of the user name itself. In this commit, the part of the code that searched for the user name was redone. Now regardless of whether the user runs with the simple command or sudo, the script searches and stores the user's name.